HashMap dividing UCs by date done
scheduleObjects ordered by date and time, hasmap that divides schedule objects with the keys being all the different dates and the values being the UC's for the dates. Tried experimenting with filtering so to make a divide and conquer approach to efficiently assign a room to a class.
Amanhã há mais zé!